format and duplication
