+++ api/src/queries/project/project-view-update-queries.ts
@@ -33,3 +33,65 @@ export const getStakeholderPartnershipsByProjectSQL = (projectId: number): SQLSt
 
   return sqlStatement;
 };
+
+/**
+ * SQL query to get project focal species.
+ *
+ * @param {number} projectId
+ * @returns {SQLStatement} sql query object
+ */
+export const getFocalSpeciesByProjectSQL = (projectId: number): SQLStatement | null => {
+  defaultLog.debug({ label: 'getFocalSpeciesByProjectSQL', message: 'params', projectId });
+
+  if (!projectId) {
+    return null;
+  }
+
+  const sqlStatement = SQL`
+    SELECT
+      name
+    FROM
+      focal_species
+    where p_id = ${projectId};
+  `;
+
+  defaultLog.debug({
+    label: 'getFocalSpeciesByProjectSQL',
+    message: 'sql',
+    'sqlStatement.text': sqlStatement.text,
+    'sqlStatement.values': sqlStatement.values
+  });
+
+  return sqlStatement;
+};
+
+/**
+ * SQL query to get project ancillary species.
+ *
+ * @param {number} projectId
+ * @returns {SQLStatement} sql query object
+ */
+export const getAncillarySpeciesByProjectSQL = (projectId: number): SQLStatement | null => {
+  defaultLog.debug({ label: 'getAncillarySpeciesByProjectSQL', message: 'params', projectId });
+
+  if (!projectId) {
+    return null;
+  }
+
+  const sqlStatement = SQL`
+    SELECT
+      name
+    FROM
+      ancillary_species
+    where p_id = ${projectId};
+  `;
+
+  defaultLog.debug({
+    label: 'getAncillarySpeciesByProjectSQL',
+    message: 'sql',
+    'sqlStatement.text': sqlStatement.text,
+    'sqlStatement.values': sqlStatement.values
+  });
+
+  return sqlStatement;
+};
